improved logic
