--- Rentit-frontend/src/authContext.js
@@ -1,20 +1,0 @@
-import React, { useState, useContext } from 'react';
-
-const AppContext = React.createContext();
-
-//Defining provider function to provide data to any components wrapped inside it
-const AppProvider = ({ children }) => {
-  const [logIn, setLogIn] = useState(false);
-  const [loginObj, setLoginObj] = useState(null);
-  // if (sessionStorage.getItem('userDetails')) setLogIn(true);
-  return (
-    <AppContext.Provider value={{ logIn, setLogIn, loginObj, setLoginObj }}>
-      {children}
-    </AppContext.Provider>
-  );
-};
-
-export const useGlobalContext = () => {
-  return useContext(AppContext);
-};
-export { AppContext, AppProvider };
