+++ Rentit-frontend/src/index.js
@@ -8,7 +8,7 @@ import Rent from './pages/rent/rent';
 import Login from './pages/login/login';
 import ItemDetails from './pages/itemDetail/itemDetails';
 // import Profile from './pages/profile/profile';
-import { AppProvider } from './authContext';
+import { AppProvider } from './Context';
 import Navbar from './components/navbar/navbar';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
