+++ Rentit-frontend/src/pages/userDetails/userDetails.css
@@ -92,9 +92,7 @@
 .productBox:active {
   transform: scale(0.9);
 }
-.rise-shake:hover {
-  animation: jump-shaking 0.83s infinite;
-}
+
 .gentle-hover-shake:hover {
   animation: tilt-shaking 0.25s infinite;
 }
@@ -184,29 +182,6 @@
   display: flex;
   align-items: center;
   justify-content: center;
-}
-@keyframes jump-shaking {
-  0% {
-    transform: translateX(0);
-  }
-  25% {
-    transform: translateY(-9px);
-  }
-  35% {
-    transform: translateY(-9px) rotate(17deg);
-  }
-  55% {
-    transform: translateY(-9px) rotate(-17deg);
-  }
-  65% {
-    transform: translateY(-9px) rotate(17deg);
-  }
-  75% {
-    transform: translateY(-9px) rotate(-17deg);
-  }
-  100% {
-    transform: translateY(0) rotate(0);
-  }
 }
 
 @keyframes tilt-shaking {
